start up a fake server with screen before tests start ref #3
