+++ Makefile
@@ -10,21 +10,27 @@ test: ## Run test with GOARCH=Default
 	@mkdir -p coverage/
 	go test -count=1 github.com/simelo/rextporter/test/config
 	go test -count=1 github.com/simelo/rextporter/src/client
+	screen -dm -S fakeSkycoinForIntegrationTest go run test/integration/fake_skycoin_node.go
 	go test -count=1 github.com/simelo/rextporter/test/integration
+	screen -S fakeSkycoinForIntegrationTest -X quit
 
 test-386: ## Run tests  with GOARCH=386
 	# GOARCH=386 go test ./cmd/... -timeout=5m
 	# GOARCH=386 go test ./src/... -timeout=5m
 	GOARCH=386 go test -count=1 github.com/simelo/rextporter/src/config
 	GOARCH=386 go test -count=1 github.com/simelo/rextporter/src/client
+	screen -dm -S fakeSkycoinForIntegrationTest go run test/integration/fake_skycoin_node.go
 	GOARCH=386 go test -count=1 github.com/simelo/rextporter/test/integration
+	screen -S fakeSkycoinForIntegrationTest -X quit
 
 test-amd64: ## Run tests with GOARCH=amd64
 	# GOARCH=amd64  go test ./cmd/... -timeout=5m
 	# GOARCH=amd64  go test ./src/... -timeout=5m
 	GOARCH=amd64 go test -count=1 github.com/simelo/rextporter/src/config
 	GOARCH=amd64 go test -count=1 github.com/simelo/rextporter/src/client
+	screen -dm -S fakeSkycoinForIntegrationTest go run test/integration/fake_skycoin_node.go
 	GOARCH=amd64 go test -count=1 github.com/simelo/rextporter/test/integration
+	screen -S fakeSkycoinForIntegrationTest -X quit
 
 lint: ## Run linters. Use make install-linters first.
 	vendorcheck ./...
